+++ src/main/kotlin/org/lgm/jobboard/jobposting/application/port/JobPostingQueryPort.kt
@@ -1,6 +1,6 @@
 package org.lgm.jobboard.jobposting.application.port
 
-import org.lgm.jobboard.jobposting.application.dto.JobPostingDetailView
+import org.lgm.jobboard.jobposting.application.query.JobPostingDetailView
 import org.lgm.jobboard.jobposting.application.query.JobPostingListItemView
 import org.lgm.jobboard.jobposting.application.query.JobPostingSearchCondition
 import org.springframework.data.domain.Page
