+++ src/main/kotlin/org/lgm/jobboard/jobposting/application/service/JobPostingQueryService.kt
@@ -1,7 +1,7 @@
 package org.lgm.jobboard.jobposting.application.service
 
-import org.lgm.jobboard.jobposting.application.dto.JobPostingDetailView
 import org.lgm.jobboard.jobposting.application.port.JobPostingQueryPort
+import org.lgm.jobboard.jobposting.application.query.JobPostingDetailView
 import org.lgm.jobboard.jobposting.application.query.JobPostingListItemView
 import org.lgm.jobboard.jobposting.application.query.JobPostingSearchCondition
 import org.lgm.jobboard.jobposting.application.query.PageView
